renamed the package name to the app name and generated a instant app bundle
