+++ app/src/androidTest/java/inc/smart/solutions/dismas/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package inc.smart.solutions.imayaprofile;
+package inc.smart.solutions.dismas;
 
 import android.content.Context;
