+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:dist="http://schemas.android.com/apk/distribution"
-    package="inc.smart.solutions.imayaprofile">
+    package="inc.smart.solutions.dismas">
 
     <dist:module dist:instant="true" />
