+++ app/src/main/java/inc/smart/solutions/dismas/models/Projects.java
@@ -1,4 +1,4 @@
-package inc.smart.solutions.imayaprofile.models;
+package inc.smart.solutions.dismas.models;
 
 public class Projects {
