+++ app/src/test/java/inc/smart/solutions/dismas/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package inc.smart.solutions.imayaprofile;
+package inc.smart.solutions.dismas;
 
 import org.junit.Test;
